Refactor: Resources from Integer to Double

+++ src/main/java/de/uniks/stp24/component/game/BuildingPropertiesComponent.java
@@ -95,7 +95,7 @@ public class BuildingPropertiesComponent extends AnchorPane {
     Map<String, String> buildingsMap;
     String buildingType;
     InGameController inGameController;
-    Map<String, Integer> priceOfBuilding;
+    Map<String, Double> priceOfBuilding;
     ObservableList<Jobs.Job> buildingJobs;
 
     Jobs.Job currentJob;
@@ -285,15 +285,15 @@ public class BuildingPropertiesComponent extends AnchorPane {
     //Sets upkeep, production and cost of buildings in listviews
     private void resourceListGeneration(BuildingAttributes buildingDto) {
         if (Objects.nonNull(buildingDto)) {
-            Map<String, Integer> resourceMapUpkeep = buildingDto.upkeep();
+            Map<String, Double> resourceMapUpkeep = buildingDto.upkeep();
             ObservableList<Resource> resourceListUpkeep = resourcesService.generateResourceList(resourceMapUpkeep, buildingConsumesListView.getItems(), null, false);
             buildingConsumesListView.setItems(resourceListUpkeep);
 
-            Map<String, Integer> resourceMapProduce = buildingDto.production();
+            Map<String, Double> resourceMapProduce = buildingDto.production();
             ObservableList<Resource> resourceListProduce = resourcesService.generateResourceList(resourceMapProduce, buildingProducesListView.getItems(), null, false);
             buildingProducesListView.setItems(resourceListProduce);
 
-            Map<String, Integer> resourceMapCost = buildingDto.cost();
+            Map<String, Double> resourceMapCost = buildingDto.cost();
             ObservableList<Resource> resourceListCost = resourcesService.generateResourceList(resourceMapCost, buildingCostsListView.getItems(), null, false);
             buildingCostsListView.setItems(resourceListCost);
