+++ src/main/java/de/uniks/stp24/dto/BuildingDto.java
@@ -5,7 +5,7 @@ import java.util.Map;
 public record BuildingDto(
         String id,
         double build_time,
-        Map<String, Integer> cost,
-        Map<String, Integer> upkeep,
-        Map<String, Integer> production) {
+        Map<String, Double> cost,
+        Map<String, Double> upkeep,
+        Map<String, Double> production) {
 }
